refactor: move UnstyledInput into SharedComponents
